Update the interface for data access of EnterExpense to fetch the current username and the monthlySpending by the username. Also added a function that checks for existence for convenience.
Reworked the EnterExpenseInteractor to calculate validity of inputs on its own (no need to access database for this).
Allowed for appending to existing records without overwriting (first fetch and only create a new one if not fetched).

Yet to implement the actual input validity testers, and to rework the EnterExpenseOutputData to be more concise.